requirements.txt
discord.py[voice]==1.7.2
aiohttp<3.7.0,>=3.6.0
async_timeout>=3.0.0
beautifulsoup4>=4.6.3
colorgram.py==1.1.0
gTTS>=2.2.1
gTTS-token>=1.1.4
SQLAlchemy==1.2.11
Pillow>=6.2.0
youtube-dl>=2018.8.28
dotabase>=7.2.3
praw==6.0.0
numpy==1.15.1
databases==0.2.2
aiosqlite==0.10.0
feedparser >= 6.0.2
python-dateutil>=2.8.1
pytest==6.2.4
pytest-asyncio==0.15.1
mock==4.0.3
httpx==0.18.1